sof-skl_hda: add priorities to use cases
Playback priorities help in Pulseaudio automatic routing with jack
events. If priorities are not added outputs will appear but nothing is
routed, because of equal priorities. This is fine when adding outputs,
but things are more problematic when outputs are removed, because audio
is still routed to already disconnected output.

Signed-off-by: Jaska Uimonen <jaska.uimonen@linux.intel.com>