sof-skl_hda: change playback volume control
Set playback volume control to Master in speakers and headphones. This
has the effect of muting both headphones and speakers at the same time,
but it is the only way currently to get the playback mute led working as
playback mute led is bound to "Master Playback Switch/Volume" in kernel
hda quirks.

Signed-off-by: Jaska Uimonen <jaska.uimonen@linux.intel.com>